requirements.txt
click==6.6
Flask==0.11.1
gitdb==0.6.4
GitPython==2.0.8
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mock==2.0.0
pbr==1.10.0
requests==2.11.1
six==1.10.0
smmap==0.9.0
Werkzeug==0.11.11
